// src/db/models/stripeConnect/products.ts
import mongoose from "mongoose";

const ObjectId = mongoose.Schema.Types.ObjectId;

export interface IProducts {
  _id: mongoose.Types.ObjectId;
  title: string;
  price: number;
  userRef: mongoose.Types.ObjectId;
  companyRef: mongoose.Types.ObjectId;
  stripeAccountId: string;
}

export interface IProductsDocument extends IProducts, mongoose.Document {
  createdAt: Date;
  updatedAt: Date;
}

// ---------------------------------------------
// This schema is part of `stripe connect` flow
// ---------------------------------------------
export const ProductSchema = new mongoose.Schema<IProductsDocument>(
  {
    title: {
      type: String,
      required: true,
    },
    price: {
      type: Number,
      required: true,
    },
    userRef: {
      type: ObjectId,
      required: true,
      ref: "User",
    },
    companyRef: {
      type: ObjectId,
      required: true,
      ref: "Company",
    },
    // Unique identifier for the connected Stripe account of the vendor.
    // This ID is generated by Stripe when a vendor successfully onboards to Stripe Connect.
    stripeAccountId: {
      type: String,
      required: true,
    },
  },
  { timestamps: true },
);

ProductSchema.index({ name: "text" });

export const StripeConnectProducts = mongoose.model<IProductsDocument>(
  "StripeConnectProducts",
  ProductSchema,
);